\name{marketmapR-package}
\alias{marketmapR-package}
\alias{marketmapR}
\docType{package}
\title{
\packageTitle{marketmapR}
}
\description{
\packageDescription{marketmapR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{marketmapR}
\packageIndices{marketmapR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{marketmapR}

Maintainer: \packageMaintainer{marketmapR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
